requirements.txt
discord.py>=2.5.0
aiohttp>=3.8.0
aiosqlite>=0.18.0
python-dotenv>=1.0.0